+++ R/cStep1.R
@@ -1,7 +1,7 @@
-aStep1 <-
-function (yesfactor, df1, df1.ls, inner.rank, initial.sample, formula, ycol, nopl, b.d) 
+cStep1 <-
+function (yesfactor, df1, df1.ls, inner.rank, initial.sample, formula, f.e, ycol, nopl, b.d) 
 {
-     #                                    aStep1   
+     #                                    cStep1   
      # REVISE ALL THIS
      # VALUE      Produces rim for Step 1. If there are factors, selects random sets from all factor subsets and runs lm
      #            and predictor to determine set with median sum of squared errors.
@@ -12,11 +12,12 @@ function (yesfactor, df1, df1.ls, inner.rank, initial.sample, formula, ycol, nop
      #            inner.rank     Rank of lm analysis on dataset with or without factor variables, depending on yesfactor     
      #            initial.sample Number of random samples from which to take rim
      #            formula        Formula for all effects including factors and constructed variables    
+     #            f.e            Right hand side of formula for Surv function
      #            ycol           Response column number
      #            nopl           n.obs.per.level 
      #            b.d            begin.diagnose Ranges from 25 to 45
      #
-     spacehere <- "%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%      aStep1           "    
+     spacehere <- "%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%      cStep1           "    
   
                            if(b.d <=31 ){ print("",quote=FALSE);print(paste(spacehere,"Section 31",sep=" "),quote=FALSE);
                                  Hmisc::prn(yesfactor);Hmisc::prn(utils::head(df1));Hmisc::prn(df1.ls);Hmisc::prn(inner.rank);
@@ -34,8 +35,10 @@ function (yesfactor, df1, df1.ls, inner.rank, initial.sample, formula, ycol, nop
      # Set up final holding variables
      nobs <- dim(df1)[1]
      Observation <- 1:initial.sample
+     Sampleno <- 1:initial.sample
      SSE <- -99
-     sumSqError <- data.frame(Observation, SSE)
+
+     sumSqError <- data.frame(Sampleno, SSE)     # from which to select median set based on SSE
      pullN <- max(inner.rank,nopl)
 
      if(yesfactor) nlevels <- length(df1.ls)
@@ -45,7 +48,6 @@ function (yesfactor, df1, df1.ls, inner.rank, initial.sample, formula, ycol, nop
      }
      hold.cands <- matrix(-9L, nrow=initial.sample, ncol=pullN*nlevels)      # will be appended when obs identified by stratum 
      for(i in 1:initial.sample){
-#prn(i)
           hold.cands.this.subset <- NULL
           for(j in 1:nlevels){
                this.subset <- df1.ls[[j]]
@@ -53,26 +55,19 @@ function (yesfactor, df1, df1.ls, inner.rank, initial.sample, formula, ycol, nop
                popu <- 1:n.in.subset
                hold.cands.temp <- sample(x=popu, size=pullN, replace=FALSE)                      # row numbers
                hold.cands.temp <- this.subset$Observation[hold.cands.temp]          # associated observation numbers
-#prn(hold.cands.this.subset)
-#prn(hold.cands.temp)
-#stop("i j")
                hold.cands.this.subset <- c(hold.cands.this.subset, hold.cands.temp)      # append observation numbers to vector
           }    #   j
           len.this.subset <- length(hold.cands.this.subset)
           hold.cands[i,1:len.this.subset] <- sort(hold.cands.this.subset) 
-#prn(hold.cands)
-#stop("i")
      }     # i
-#prn(hold.cands)
-#stop("after loop i")
 
                            if(b.d <=32 ){ print("",quote=FALSE);print(paste(spacehere,"Section 32",sep=" "),quote=FALSE);
                                  Hmisc::prn(inner.rank);Hmisc::prn(nopl);Hmisc::prn(nobs);Hmisc::prn(utils::head(sumSqError));Hmisc::prn(nlevels);
-                                 Hmisc::prn(hold.cands);Hmisc::prn(0);Hmisc::prn(0);Hmisc::prn(0)   }
+                                 Hmisc::prn(utils::head(hold.cands));Hmisc::prn(utils::tail(hold.cands))   }
 
 
      ################################################################################################# 
-     # Run the lm function on each row of hold.cands and predict to the entire database for each one #
+     # Run the coxph function on each row of hold.cands and predict to the entire database for each one #
      # Calculate the sum of squares of the error for each one amd store it in sumSqError[i,2]        #
      # Sort this matrix on the 2nd column and locate the median row.  The observations for this row  #
      # are the rim for Step 1                                                                        #
@@ -81,28 +76,28 @@ function (yesfactor, df1, df1.ls, inner.rank, initial.sample, formula, ycol, nop
      for(i in 1:initial.sample){
           index <- hold.cands[i,]
           smalldata <- df1[index,]
-          this.form <- formula
 
-          lmsmall <- stats::lm(formula=this.form, data=smalldata, singular.ok=TRUE)                             #    lm
-          predsmall <- stats::predict(lmsmall, data=df1, type="response", pred.var=1)                          # predict
+          thisform <- paste("survival::Surv(time=event.time, event=status)", f.e, sep=" ~ ")
+          thisform <- stats::as.formula(thisform)
+
+          lmsmall <- do.call(what=survival::coxph, args=list(formula=thisform, data=smalldata, 
+               ties="efron", model=TRUE, singular.ok=FALSE, x=TRUE, y=TRUE))                                      # coxph
+
+          predsmall <- stats::predict(object=lmsmall)                                                             #  predict
 
           errorsmall <- df1[, ycol] - predsmall
-          sserrorsmall <- errorsmall^2
-          sumSqError[i,2] <- sum(sserrorsmall)
+          predsmall2 <- predsmall^2
+          sumSqError[i,2] <- sum(predsmall2)
      }     # i
-#prn(sumSqError)
-
      sumSqError <- sumSqError[order(sumSqError[,2]),]
-#prn(sumSqError)
-     MED <- round(initial.sample/2)
-#prn(MED)
-#prn(hold.cands)
+     hold.cands <- hold.cands[sumSqError[,1],]
+     MED <- round(initial.sample/2 + .000001)
+
                            if(b.d <=34 ){ print("",quote=FALSE);print(paste(spacehere,"Section 34",sep=" "),quote=FALSE);
-                                 Hmisc::prn(hold.cands);Hmisc::prn(smalldata);Hmisc::prn(this.form);Hmisc::prn(predsmall);Hmisc::prn(sumSqError);
-                                 Hmisc::prn(MED);Hmisc::prn(hold.cands)    }
+                                 Hmisc::prn(smalldata);Hmisc::prn(thisform);Hmisc::prn(predsmall);Hmisc::prn(utils::head(sumSqError));
+                                 Hmisc::prn(utils::tail(sumSqError))
+                                 Hmisc::prn(MED);Hmisc::prn(utils::head(hold.cands));Hmisc::prn(utils::tail(hold.cands))    }
 
      rimout <- hold.cands[MED,]
-#prn(rimout)
-#stop("step 1")
      return(rimout)
 }